Time: 3 ms (33.84%), Space: 11.4 MB (12.07%) - LeetHub

+++ 1910-remove-all-occurrences-of-a-substring/1910-remove-all-occurrences-of-a-substring.cpp
@@ -3,7 +3,7 @@ public:
     string removeOccurrences(string s, string part) {
         if (s.find(part) == string::npos)
             return s;
-        size_t pos = s.find(part);
+        int pos = s.find(part);
         if (pos != string::npos) {
             s.erase(pos, part.size());
